isolate asset syncing, to avoid syncing twice
